requirements.txt
ansible==9.2.0
faiss-cpu==1.7.4
langchain==0.1.6
langchain-openai==0.0.5
Markdown==3.5.2
netmiko==4.3.0
pillow==10.2.0
streamlit==1.31.1
unstructured==0.12.4
watchdog==4.0.0
